# src/openapi.yaml
openapi: 3.0.3
info:
  title: DIO Bank
  description: DIO Bank API
  version: 1.0.0

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      required:
        - username
        - password
        - role_id
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: string
        role_id:
          type: integer

    Role:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
        name:
          type: string

    Auth:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

security:
  - bearerAuth: []

paths:
  /users:
    get:
      summary: List all users
      tags: [user]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create a new user
      tags: [user]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Validation error

  /users/{user_id}:
    get:
      summary: Get user by ID
      tags: [user]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      summary: Update user by ID
      tags: [user]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
        '404':
          description: User not found

    delete:
      summary: Delete user by ID
      tags: [user]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found

  /roles:
    get:
      summary: List all roles
      tags: [role]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'

    post:
      summary: Create a new role
      tags: [role]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: Role created successfully

  /auth/login:
    post:
      summary: Authenticate user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials